utils: add stateless retracer utility
The retracer utility reads and replays a json-formatted file produced by
the stateless tracer utility. The retracer runs independently from other
userspace applications and so allows a driver to be tested in different
userspace environments. If the trace file is edited to send specific
errors to a driver, the retracer can be used to test a driver's error-
handling abilities.

Signed-off-by: Deborah Brouwer <deborah.brouwer@collabora.com>